+++ src/Yasumi/data/translations/assumptionOfMary.php
@@ -12,6 +12,7 @@
 
 // Translations for Assumption of Mary
 return [
+    'ca' => 'l\'Assumpció',
     'de' => 'Mariä Himmelfahrt',
     'el' => 'Κοίμηση της Θεοτόκου',
     'en' => 'Assumption of Mary',