+++ src/Yasumi/data/translations/christmasDay.php
@@ -13,6 +13,7 @@
 // Translations for Christmas
 return [
     'bs_Latn' => 'Božić',
+    'ca' => 'Nadal',
     'cs' => '1. svátek vánoční',
     'cy' => 'Nadolig',
     'da' => 'Juledag',