+++ src/Yasumi/data/translations/easter.php
@@ -13,6 +13,7 @@
 // Translations for Easter Sunday
 return [
     'bs_Latn' => 'Uskrs',
+    'ca' => 'Pasqua',
     'cy' => 'Sul y Pasg',
     'da' => 'Påskedag',
     'de' => 'Ostersonntag',