+++ src/Yasumi/data/translations/epiphany.php
@@ -12,6 +12,7 @@
 
 // Translations for Epiphany
 return [
+    'ca' => 'Epifania',
     'de_AT' => 'Heilige Drei Könige',
     'de_CH' => 'Heilige Drei Könige',
     'de' => 'Heilige 3 Könige',