+++ src/Yasumi/data/translations/immaculateConception.php
@@ -12,6 +12,7 @@
 
 // Translations for Immaculate Conception
 return [
+    'ca' => 'Immaculada Concepció',
     'de' => 'Mariä Empfängnis',
     'el' => 'Ευαγγελισμός της Θεοτόκου',
     'en' => 'Immaculate Conception',