+++ src/Yasumi/data/translations/maundyThursday.php
@@ -12,6 +12,7 @@
 
 // Translations for Maundy Thursday
 return [
+    'ca' => 'dijous Sant',
     'da' => 'Skærtorsdag',
     'el' => 'Μεγάλη Πέμπτη',
     'en' => 'Maundy Thursday',