+++ src/Yasumi/data/translations/valentinesDay.php
@@ -12,6 +12,7 @@
 
 // Translations for Valentine's Day
 return [
+    'ca' => 'Dia de Sant Valentí',
     'de' => 'Valentinstag',
     'el' => 'Αγίου Βαλεντίνου',
     'en' => 'Valentine\'s Day',